add multimap

+++ code/CCoA_ij_idea/src/Logic/Landscape/ZonaLandscape.java
@@ -4,5 +4,5 @@ import Logic.FootprintSpaceTime.PolygonExtended;
 import Logic.TypesInLevel;
 
 public interface ZonaLandscape {
-    public boolean getAccessPlace(PolygonExtended place, int time, TypesInLevel type);
+    public boolean getAccessPlace(PolygonExtended place, double time, TypesInLevel type);
 }
